Reusing functions using common header
